+++ metric_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"go.opentelemetry.io/otel/metric"
+	"go.opentelemetry.io/otel/metric/noop"
 )
 
 type testMeterProvider struct{}
@@ -27,12 +28,12 @@ type testMeterProvider struct{}
 var _ metric.MeterProvider = &testMeterProvider{}
 
 func (*testMeterProvider) Meter(_ string, _ ...metric.MeterOption) metric.Meter {
-	return metric.NewNoopMeterProvider().Meter("")
+	return noop.NewMeterProvider().Meter("")
 }
 
 func TestMultipleGlobalMeterProvider(t *testing.T) {
 	p1 := testMeterProvider{}
-	p2 := metric.NewNoopMeterProvider()
+	p2 := noop.NewMeterProvider()
 	SetMeterProvider(&p1)
 	SetMeterProvider(p2)